+++ limbo/library/src/main/java/co/windly/limbo/activity/lce/LimboLceActivityView.kt
@@ -1,6 +1,0 @@
-package co.windly.limbo.activity.lce
-
-import co.windly.limbo.LimboView
-import com.hannesdorfmann.mosby3.mvp.lce.MvpLceView
-
-interface LimboLceActivityView<M> : LimboView, MvpLceView<M>