+++ limbo/mvp/src/main/java/co/windly/limbo/mvp/LimboView.kt
@@ -1,0 +1,18 @@
+package co.windly.limbo.mvp
+
+import com.hannesdorfmann.mosby3.mvp.MvpView
+import io.reactivex.disposables.CompositeDisposable
+import io.reactivex.disposables.Disposable
+
+interface LimboView : MvpView {
+
+  //region Reactive
+
+  val disposables: CompositeDisposable
+
+  fun addDisposable(disposable: Disposable): Boolean
+
+  fun clearDisposables()
+
+  //endregion
+}