+++ samples/java/src/main/java/co/windly/limbosample/presentation/main/root/RootPresenter.java
@@ -1,6 +1,6 @@
 package co.windly.limbosample.presentation.main.root;
 
-import co.windly.limbo.presenter.queue.LimboQueuePresenter;
+import co.windly.limbo.mvp.presenter.queue.LimboQueuePresenter;
 import javax.inject.Inject;
 
 class RootPresenter extends LimboQueuePresenter<RootView> {