+++ samples/kotlin/src/main/java/co/windly/limbosample/presentation/main/MainPresenter.kt
@@ -1,6 +1,6 @@
 package co.windly.limbosample.presentation.main
 
-import co.windly.limbo.presenter.queue.LimboQueuePresenter
+import co.windly.limbo.mvp.presenter.queue.LimboQueuePresenter
 import javax.inject.Inject
 
 class MainPresenter @Inject constructor() : LimboQueuePresenter<MainView>()