+++ samples/kotlin/src/main/java/co/windly/limbosample/presentation/main/root/home/HomeView.kt
@@ -1,6 +1,6 @@
 package co.windly.limbosample.presentation.main.root.home
 
-import co.windly.limbo.fragment.base.LimboFragmentView
+import co.windly.limbo.mvp.fragment.base.LimboFragmentView
 
 interface HomeView : LimboFragmentView {
   // No-op.