+++ samples/kotlin/src/main/java/co/windly/limbosample/presentation/splash/SplashView.kt
@@ -1,6 +1,6 @@
 package co.windly.limbosample.presentation.splash
 
-import co.windly.limbo.fragment.base.LimboFragmentView
+import co.windly.limbo.mvp.fragment.base.LimboFragmentView
 
 interface SplashView : LimboFragmentView {